autopmo-api/requirements.txt
azure-core==1.35.0
azure-functions==1.23.0
azure-storage-blob==12.26.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
idna==3.10
isodate==0.7.2
lxml==6.0.0
MarkupSafe==3.0.2
pycparser==2.22
python-docx==1.2.0
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
